create 3 child components
