+++ plainly-plugin/src/ui/components/layers/index.ts
@@ -1,3 +1,6 @@
+export * from './BulkScriptSelect';
 export * from './CropScriptDialog';
+export * from './FilterLayers';
 export * from './Layers';
+export * from './ParametrizedLayers';
 export * from './ScriptsDialog';
